Changed input number type to float

+++ Math_calc.py
@@ -2,7 +2,7 @@ import math
 
 
 print("Hi. This is simple calculation software.")
-x = int(input("Please, indicate your number : "))
+x = float(input("Please, indicate your number : "))
 y = int(input("""Please specify what do you want to do with the number : ")
     1. Calculate sin
     2. Calculate cos
